Make ProgressBar and ProgressbarItem web components
ProgressBar and ProgressBarItem are rewritten as vanilla web components,
while ProficienyBar is removed (it is just a use case for ProgressBar).

ProgressBar stories use the storyContext.globals.theme
for theming.

The components are inspired by shoelace and fast. But as shoelace and fast
only show one progress bar, we need to be able to show two.
Also, neither of these support using aria-valuetext to explain the value.

The react version took an array of progress bar properties as a prop. As web
components' attributes should not be arrays or objects, the web component
version rather take ProgressBarItems as children.

Some conventions that make sense:
- pass around full css values (not just custom property names).
  Full css values work in any case, while custom property names work only
  if they are defined elsewhere.
- put all --kobber references in constructor, to enable a simple overview.