Enable kobber-button to be used together with other interactives
This makes kobber-button respect prop usedInOtherInteractive,
that implements some own styling and functionality.

For example, if one wants to use kobber-button as kobber-radio-input,
kobber-buttons normal height constraint should not apply. And
kobber-radio-input has its own tabindex, so double focus is avoided.

This is not how buttons should behave normally, therefore this is
not documented in kobber-button story.