Add radio react-ssr components
This includes three things:

1. Adding properties name and  value to kobber-button, which is required
by ShoeLace. For some reason, these need to be added to the web component
version of buttons after adding radio react files.
2. Changing an internal shoelace file, by removing the string type
specification for value in
https://github.com/shoelace-style/shoelace/blob/next/src/internal/form.ts#L91
Without this removal, build fails with error TS2322, saying "Type
'unknown' is not assignable to type 'string'."
3. Avoid using checked-icon in react (due to icons not being ssr-safe),
but rather setting background-color to black.