Preliminarily correct radio font usage
This ensures radio input actually uses the PP Mori font. And it corrects
the input control being seemingly too low compared to its label.

Preliminarily, because font setting must be handled better. Per today
this only works in docs and storybook, not in the wild (because of the
PP Mori font definitions).