Reuse existing stories across stories
Instead of hard coding copies of those stories, and their controls
(which is error-prone, and hard to maintain).

This requires us to be stricter to put decorator elements in stories'
decorators array, so they don't leak into stories reusing the stories
(which will show reused stories wrong, and confuse consumers).